task 2 - CNN implementation/requirements.txt
torch
torchvision
matplotlib
scikit-learn
tqdm
Pillow